Add helpers/config.js to generate compiled configuration for application
